+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Ability to add, edit and delete notes inside a Deck.
 - Implementation for study section with all Decks with pending flashcards.
 - Study page.
+- Text alignment controls in editor.
 
 ### Changed
 - Disabled React's Concurrent Mode.